Add cmake flag for iceberg native connector
