+++ presto-native-execution/presto_cpp/main/connectors/PrestoToVeloxConnector.cpp
@@ -16,7 +16,6 @@
 #include "presto_cpp/main/types/PrestoToVeloxExpr.h"
 #include "presto_cpp/main/types/TypeParser.h"
 #include "presto_cpp/presto_protocol/connector/hive/HiveConnectorProtocol.h"
-#include "presto_cpp/presto_protocol/connector/iceberg/IcebergConnectorProtocol.h"
 #include "presto_cpp/presto_protocol/connector/tpch/TpchConnectorProtocol.h"
 
 #include <velox/type/fbhive/HiveTypeParser.h>
@@ -24,8 +23,6 @@
 #include "velox/connectors/hive/HiveConnectorSplit.h"
 #include "velox/connectors/hive/HiveDataSink.h"
 #include "velox/connectors/hive/TableHandle.h"
-#include "velox/connectors/hive/iceberg/IcebergDeleteFile.h"
-#include "velox/connectors/hive/iceberg/IcebergSplit.h"
 #include "velox/connectors/tpch/TpchConnector.h"
 #include "velox/connectors/tpch/TpchConnectorSplit.h"
 #include "velox/type/Filter.h"
@@ -33,6 +30,8 @@
 namespace facebook::presto {
 
 namespace {
+using namespace facebook::velox;
+
 std::unordered_map<std::string, std::unique_ptr<const PrestoToVeloxConnector>>&
 connectors() {
   static std::
@@ -42,30 +41,6 @@ connectors() {
 }
 } // namespace
 
-void registerPrestoToVeloxConnector(
-    std::unique_ptr<const PrestoToVeloxConnector> connector) {
-  auto connectorName = connector->connectorName();
-  auto connectorProtocol = connector->createConnectorProtocol();
-  VELOX_CHECK(
-      connectors().insert({connectorName, std::move(connector)}).second,
-      "Connector {} is already registered",
-      connectorName);
-  protocol::registerConnectorProtocol(
-      connectorName, std::move(connectorProtocol));
-}
-
-void unregisterPrestoToVeloxConnector(const std::string& connectorName) {
-  connectors().erase(connectorName);
-  protocol::unregisterConnectorProtocol(connectorName);
-}
-
-const PrestoToVeloxConnector& getPrestoToVeloxConnector(
-    const std::string& connectorName) {
-  auto it = connectors().find(connectorName);
-  VELOX_CHECK(
-      it != connectors().end(), "Connector {} not registered", connectorName);
-  return *(it->second);
-}
 
 namespace {
 using namespace velox;
@@ -102,50 +77,9 @@ dwio::common::FileFormat toVeloxFileFormat(
       "Unsupported file format: {} {}", format.inputFormat, format.serDe);
 }
 
-dwio::common::FileFormat toVeloxFileFormat(
-    const presto::protocol::iceberg::FileFormat format) {
-  if (format == protocol::iceberg::FileFormat::ORC) {
-    return dwio::common::FileFormat::ORC;
-  } else if (format == protocol::iceberg::FileFormat::PARQUET) {
-    return dwio::common::FileFormat::PARQUET;
-  }
-  VELOX_UNSUPPORTED("Unsupported file format: {}", fmt::underlying(format));
-}
-
 template <typename T>
 std::string toJsonString(const T& value) {
   return ((json)value).dump();
-}
-
-TypePtr stringToType(
-    const std::string& typeString,
-    const TypeParser& typeParser) {
-  return typeParser.parse(typeString);
-}
-
-connector::hive::HiveColumnHandle::ColumnType toHiveColumnType(
-    protocol::hive::ColumnType type) {
-  switch (type) {
-    case protocol::hive::ColumnType::PARTITION_KEY:
-      return connector::hive::HiveColumnHandle::ColumnType::kPartitionKey;
-    case protocol::hive::ColumnType::REGULAR:
-      return connector::hive::HiveColumnHandle::ColumnType::kRegular;
-    case protocol::hive::ColumnType::SYNTHESIZED:
-      return connector::hive::HiveColumnHandle::ColumnType::kSynthesized;
-    default:
-      VELOX_UNSUPPORTED(
-          "Unsupported Hive column type: {}.", toJsonString(type));
-  }
-}
-
-std::vector<common::Subfield> toRequiredSubfields(
-    const protocol::List<protocol::Subfield>& subfields) {
-  std::vector<common::Subfield> result;
-  result.reserve(subfields.size());
-  for (auto& subfield : subfields) {
-    result.emplace_back(subfield);
-  }
-  return result;
 }
 
 template <TypeKind KIND>
@@ -800,81 +734,6 @@ std::unique_ptr<common::Filter> toFilter(
   VELOX_UNSUPPORTED("Unsupported filter found.");
 }
 
-std::unique_ptr<connector::ConnectorTableHandle> toHiveTableHandle(
-    const protocol::TupleDomain<protocol::Subfield>& domainPredicate,
-    const std::shared_ptr<protocol::RowExpression>& remainingPredicate,
-    bool isPushdownFilterEnabled,
-    const std::string& tableName,
-    const protocol::List<protocol::Column>& dataColumns,
-    const protocol::TableHandle& tableHandle,
-    const protocol::Map<protocol::String, protocol::String>& tableParameters,
-    const VeloxExprConverter& exprConverter,
-    const TypeParser& typeParser) {
-  common::SubfieldFilters subfieldFilters;
-  auto domains = domainPredicate.domains;
-  for (const auto& domain : *domains) {
-    auto filter = domain.second;
-    subfieldFilters[common::Subfield(domain.first)] =
-        toFilter(domain.second, exprConverter, typeParser);
-  }
-
-  auto remainingFilter = exprConverter.toVeloxExpr(remainingPredicate);
-  if (auto constant = std::dynamic_pointer_cast<const core::ConstantTypedExpr>(
-          remainingFilter)) {
-    bool value = constant->value().value<bool>();
-    VELOX_CHECK(value, "Unexpected always-false remaining predicate");
-
-    // Use null for always-true filter.
-    remainingFilter = nullptr;
-  }
-
-  RowTypePtr finalDataColumns;
-  if (!dataColumns.empty()) {
-    std::vector<std::string> names;
-    std::vector<TypePtr> types;
-    velox::type::fbhive::HiveTypeParser hiveTypeParser;
-    names.reserve(dataColumns.size());
-    types.reserve(dataColumns.size());
-    for (auto& column : dataColumns) {
-      std::string name = column.name;
-      folly::toLowerAscii(name);
-      names.emplace_back(std::move(name));
-      auto parsedType = hiveTypeParser.parse(column.type);
-      // The type from the metastore may have upper case letters
-      // in field names, convert them all to lower case to be
-      // compatible with Presto.
-      types.push_back(VELOX_DYNAMIC_TYPE_DISPATCH(
-          fieldNamesToLowerCase, parsedType->kind(), parsedType));
-    }
-    finalDataColumns = ROW(std::move(names), std::move(types));
-  }
-
-  if (tableParameters.empty()) {
-    return std::make_unique<connector::hive::HiveTableHandle>(
-        tableHandle.connectorId,
-        tableName,
-        isPushdownFilterEnabled,
-        std::move(subfieldFilters),
-        remainingFilter,
-        finalDataColumns);
-  }
-
-  std::unordered_map<std::string, std::string> finalTableParameters = {};
-  finalTableParameters.reserve(tableParameters.size());
-  for (const auto& [key, value] : tableParameters) {
-    finalTableParameters[key] = value;
-  }
-
-  return std::make_unique<connector::hive::HiveTableHandle>(
-      tableHandle.connectorId,
-      tableName,
-      isPushdownFilterEnabled,
-      std::move(subfieldFilters),
-      remainingFilter,
-      finalDataColumns,
-      finalTableParameters);
-}
-
 connector::hive::LocationHandle::TableType toTableType(
     protocol::hive::TableType tableType) {
   switch (tableType) {
@@ -1088,17 +947,138 @@ velox::connector::hive::HiveBucketConversion toVeloxBucketConversion(
   return veloxBucketConversion;
 }
 
-velox::connector::hive::iceberg::FileContent toVeloxFileContent(
-    const presto::protocol::iceberg::FileContent content) {
-  if (content == protocol::iceberg::FileContent::DATA) {
-    return velox::connector::hive::iceberg::FileContent::kData;
-  } else if (content == protocol::iceberg::FileContent::POSITION_DELETES) {
-    return velox::connector::hive::iceberg::FileContent::kPositionalDeletes;
-  }
-  VELOX_UNSUPPORTED("Unsupported file content: {}", fmt::underlying(content));
+} // namespace
+
+void registerPrestoToVeloxConnector(
+    std::unique_ptr<const PrestoToVeloxConnector> connector) {
+  auto connectorName = connector->connectorName();
+  auto connectorProtocol = connector->createConnectorProtocol();
+  VELOX_CHECK(
+      connectors().insert({connectorName, std::move(connector)}).second,
+      "Connector {} is already registered",
+      connectorName);
+  protocol::registerConnectorProtocol(
+      connectorName, std::move(connectorProtocol));
 }
 
-} // namespace
+void unregisterPrestoToVeloxConnector(const std::string& connectorName) {
+  connectors().erase(connectorName);
+  protocol::unregisterConnectorProtocol(connectorName);
+}
+
+TypePtr stringToType(
+    const std::string& typeString,
+    const TypeParser& typeParser) {
+  return typeParser.parse(typeString);
+}
+
+std::vector<common::Subfield> toRequiredSubfields(
+    const protocol::List<protocol::Subfield>& subfields) {
+  std::vector<common::Subfield> result;
+  result.reserve(subfields.size());
+  for (auto& subfield : subfields) {
+    result.emplace_back(subfield);
+  }
+  return result;
+}
+
+const PrestoToVeloxConnector& getPrestoToVeloxConnector(
+    const std::string& connectorName) {
+  auto it = connectors().find(connectorName);
+  VELOX_CHECK(
+      it != connectors().end(), "Connector {} not registered", connectorName);
+  return *(it->second);
+}
+
+connector::hive::HiveColumnHandle::ColumnType toHiveColumnType(
+    protocol::hive::ColumnType type) {
+  switch (type) {
+    case protocol::hive::ColumnType::PARTITION_KEY:
+      return connector::hive::HiveColumnHandle::ColumnType::kPartitionKey;
+    case protocol::hive::ColumnType::REGULAR:
+      return connector::hive::HiveColumnHandle::ColumnType::kRegular;
+    case protocol::hive::ColumnType::SYNTHESIZED:
+      return connector::hive::HiveColumnHandle::ColumnType::kSynthesized;
+    default:
+      VELOX_UNSUPPORTED(
+          "Unsupported Hive column type: {}.", toJsonString(type));
+  }
+}
+
+std::unique_ptr<connector::ConnectorTableHandle> toHiveTableHandle(
+    const protocol::TupleDomain<protocol::Subfield>& domainPredicate,
+    const std::shared_ptr<protocol::RowExpression>& remainingPredicate,
+    bool isPushdownFilterEnabled,
+    const std::string& tableName,
+    const protocol::List<protocol::Column>& dataColumns,
+    const protocol::TableHandle& tableHandle,
+    const protocol::Map<protocol::String, protocol::String>& tableParameters,
+    const VeloxExprConverter& exprConverter,
+    const TypeParser& typeParser) {
+  common::SubfieldFilters subfieldFilters;
+  auto domains = domainPredicate.domains;
+  for (const auto& domain : *domains) {
+    auto filter = domain.second;
+    subfieldFilters[common::Subfield(domain.first)] =
+        toFilter(domain.second, exprConverter, typeParser);
+  }
+
+  auto remainingFilter = exprConverter.toVeloxExpr(remainingPredicate);
+  if (auto constant = std::dynamic_pointer_cast<const core::ConstantTypedExpr>(
+          remainingFilter)) {
+    bool value = constant->value().value<bool>();
+    VELOX_CHECK(value, "Unexpected always-false remaining predicate");
+
+    // Use null for always-true filter.
+    remainingFilter = nullptr;
+  }
+
+  RowTypePtr finalDataColumns;
+  if (!dataColumns.empty()) {
+    std::vector<std::string> names;
+    std::vector<TypePtr> types;
+    velox::type::fbhive::HiveTypeParser hiveTypeParser;
+    names.reserve(dataColumns.size());
+    types.reserve(dataColumns.size());
+    for (auto& column : dataColumns) {
+      std::string name = column.name;
+      folly::toLowerAscii(name);
+      names.emplace_back(std::move(name));
+      auto parsedType = hiveTypeParser.parse(column.type);
+      // The type from the metastore may have upper case letters
+      // in field names, convert them all to lower case to be
+      // compatible with Presto.
+      types.push_back(VELOX_DYNAMIC_TYPE_DISPATCH(
+          fieldNamesToLowerCase, parsedType->kind(), parsedType));
+    }
+    finalDataColumns = ROW(std::move(names), std::move(types));
+  }
+
+  if (tableParameters.empty()) {
+    return std::make_unique<connector::hive::HiveTableHandle>(
+        tableHandle.connectorId,
+        tableName,
+        isPushdownFilterEnabled,
+        std::move(subfieldFilters),
+        remainingFilter,
+        finalDataColumns);
+  }
+
+  std::unordered_map<std::string, std::string> finalTableParameters = {};
+  finalTableParameters.reserve(tableParameters.size());
+  for (const auto& [key, value] : tableParameters) {
+    finalTableParameters[key] = value;
+  }
+
+  return std::make_unique<connector::hive::HiveTableHandle>(
+      tableHandle.connectorId,
+      tableName,
+      isPushdownFilterEnabled,
+      std::move(subfieldFilters),
+      remainingFilter,
+      finalDataColumns,
+      finalTableParameters);
+}
 
 std::unique_ptr<velox::connector::ConnectorSplit>
 HivePrestoToVeloxConnector::toVeloxSplit(
@@ -1332,163 +1312,6 @@ HivePrestoToVeloxConnector::createVeloxPartitionFunctionSpec(
 std::unique_ptr<protocol::ConnectorProtocol>
 HivePrestoToVeloxConnector::createConnectorProtocol() const {
   return std::make_unique<protocol::hive::HiveConnectorProtocol>();
-}
-
-std::unique_ptr<velox::connector::ConnectorSplit>
-IcebergPrestoToVeloxConnector::toVeloxSplit(
-    const protocol::ConnectorId& catalogId,
-    const protocol::ConnectorSplit* connectorSplit,
-    const protocol::SplitContext* splitContext) const {
-  auto icebergSplit =
-      dynamic_cast<const protocol::iceberg::IcebergSplit*>(connectorSplit);
-  VELOX_CHECK_NOT_NULL(
-      icebergSplit, "Unexpected split type {}", connectorSplit->_type);
-
-  std::unordered_map<std::string, std::optional<std::string>> partitionKeys;
-  for (const auto& entry : icebergSplit->partitionKeys) {
-    partitionKeys.emplace(
-        entry.second.name,
-        entry.second.value == nullptr
-            ? std::nullopt
-            : std::optional<std::string>{*entry.second.value});
-  }
-
-  std::unordered_map<std::string, std::string> customSplitInfo;
-  customSplitInfo["table_format"] = "hive-iceberg";
-
-  std::vector<velox::connector::hive::iceberg::IcebergDeleteFile> deletes;
-  deletes.reserve(icebergSplit->deletes.size());
-  for (const auto& deleteFile : icebergSplit->deletes) {
-    std::unordered_map<int32_t, std::string> lowerBounds(
-        deleteFile.lowerBounds.begin(), deleteFile.lowerBounds.end());
-
-    std::unordered_map<int32_t, std::string> upperBounds(
-        deleteFile.upperBounds.begin(), deleteFile.upperBounds.end());
-
-    velox::connector::hive::iceberg::IcebergDeleteFile icebergDeleteFile(
-        toVeloxFileContent(deleteFile.content),
-        deleteFile.path,
-        toVeloxFileFormat(deleteFile.format),
-        deleteFile.recordCount,
-        deleteFile.fileSizeInBytes,
-        std::vector(deleteFile.equalityFieldIds),
-        lowerBounds,
-        upperBounds);
-
-    deletes.emplace_back(icebergDeleteFile);
-  }
-
-  std::unordered_map<std::string, std::string> infoColumns = {
-      {"$data_sequence_number",
-       std::to_string(icebergSplit->dataSequenceNumber)},
-      {"$path", icebergSplit->path}};
-
-  return std::make_unique<connector::hive::iceberg::HiveIcebergSplit>(
-      catalogId,
-      icebergSplit->path,
-      toVeloxFileFormat(icebergSplit->fileFormat),
-      icebergSplit->start,
-      icebergSplit->length,
-      partitionKeys,
-      std::nullopt,
-      customSplitInfo,
-      nullptr,
-      splitContext->cacheable,
-      deletes,
-      infoColumns);
-}
-
-std::unique_ptr<velox::connector::ColumnHandle>
-IcebergPrestoToVeloxConnector::toVeloxColumnHandle(
-    const protocol::ColumnHandle* column,
-    const TypeParser& typeParser) const {
-  auto icebergColumn =
-      dynamic_cast<const protocol::iceberg::IcebergColumnHandle*>(column);
-  VELOX_CHECK_NOT_NULL(
-      icebergColumn, "Unexpected column handle type {}", column->_type);
-  // TODO(imjalpreet): Modify 'hiveType' argument of the 'HiveColumnHandle'
-  //  constructor similar to how Hive Connector is handling for bucketing
-  velox::type::fbhive::HiveTypeParser hiveTypeParser;
-  auto type = stringToType(icebergColumn->type, typeParser);
-  connector::hive::HiveColumnHandle::ColumnParseParameters columnParseParameters;
-  if (type->isDate()) {
-    columnParseParameters.partitionDateValueFormat = connector::hive::HiveColumnHandle::ColumnParseParameters::kDaysSinceEpoch;
-  }
-  return std::make_unique<connector::hive::HiveColumnHandle>(
-      icebergColumn->columnIdentity.name,
-      toHiveColumnType(icebergColumn->columnType),
-      type,
-      type,
-      toRequiredSubfields(icebergColumn->requiredSubfields),
-      columnParseParameters);
-}
-
-std::unique_ptr<velox::connector::ConnectorTableHandle>
-IcebergPrestoToVeloxConnector::toVeloxTableHandle(
-    const protocol::TableHandle& tableHandle,
-    const VeloxExprConverter& exprConverter,
-    const TypeParser& typeParser,
-    velox::connector::ColumnHandleMap& assignments) const {
-  auto addSynthesizedColumn = [&](const std::string& name,
-                                  protocol::hive::ColumnType columnType,
-                                  const protocol::ColumnHandle& column) {
-    if (toHiveColumnType(columnType) ==
-        velox::connector::hive::HiveColumnHandle::ColumnType::kSynthesized) {
-      if (assignments.count(name) == 0) {
-        assignments.emplace(name, toVeloxColumnHandle(&column, typeParser));
-      }
-    }
-  };
-
-  auto icebergLayout = std::dynamic_pointer_cast<
-      const protocol::iceberg::IcebergTableLayoutHandle>(
-      tableHandle.connectorTableLayout);
-  VELOX_CHECK_NOT_NULL(
-      icebergLayout,
-      "Unexpected layout type {}",
-      tableHandle.connectorTableLayout->_type);
-
-  for (const auto& entry : icebergLayout->partitionColumns) {
-    assignments.emplace(
-        entry.columnIdentity.name, toVeloxColumnHandle(&entry, typeParser));
-  }
-
-  // Add synthesized columns to the TableScanNode columnHandles as well.
-  for (const auto& entry : icebergLayout->predicateColumns) {
-    addSynthesizedColumn(entry.first, entry.second.columnType, entry.second);
-  }
-
-  auto icebergTableHandle =
-      std::dynamic_pointer_cast<const protocol::iceberg::IcebergTableHandle>(
-          tableHandle.connectorHandle);
-  VELOX_CHECK_NOT_NULL(
-      icebergTableHandle,
-      "Unexpected table handle type {}",
-      tableHandle.connectorHandle->_type);
-
-  // Use fully qualified name if available.
-  std::string tableName = icebergTableHandle->schemaName.empty()
-      ? icebergTableHandle->icebergTableName.tableName
-      : fmt::format(
-            "{}.{}",
-            icebergTableHandle->schemaName,
-            icebergTableHandle->icebergTableName.tableName);
-
-  return toHiveTableHandle(
-      icebergLayout->domainPredicate,
-      icebergLayout->remainingPredicate,
-      icebergLayout->pushdownFilterEnabled,
-      tableName,
-      icebergLayout->dataColumns,
-      tableHandle,
-      {},
-      exprConverter,
-      typeParser);
-}
-
-std::unique_ptr<protocol::ConnectorProtocol>
-IcebergPrestoToVeloxConnector::createConnectorProtocol() const {
-  return std::make_unique<protocol::iceberg::IcebergConnectorProtocol>();
 }
 
 std::unique_ptr<velox::connector::ConnectorSplit>
