+++ presto-native-execution/presto_cpp/main/types/PrestoToVeloxQueryPlan.cpp
@@ -41,12 +41,6 @@ namespace facebook::presto {
 
 namespace {
 
-TypePtr stringToType(
-    const std::string& typeString,
-    const TypeParser& typeParser) {
-  return typeParser.parse(typeString);
-}
-
 std::vector<std::string> getNames(const protocol::Assignments& assignments) {
   std::vector<std::string> names;
   names.reserve(assignments.assignments.size());
